pyproject.toml
[tool.poetry]
name = "ms authentication"
version = "0.1.0"
description = ""
authors = ["JonasR <email>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-decouple = "^3.8"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
sqlalchemy-utils = "^0.41.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
